Add framework for tested examples
Examples are just .t files that run like any other test.  They are
marked up with special comments:

The tests include a call to a subroutine to process the test file into
markdown.  That markdown file lives in the docs/examples folder and
should be included in the mkdocs.yml in order to show up on readthedocs.

To preview the docs:

$ sudo pip install mkdocs
$ mkdocs serve -a 0.0.0.0:8000

then on your host machine go to http://192.168.20.20:8000